cli: Add `--max-line-length` to exclude long lines
Sometimes search results include very long lines, for example
concatenated JavaScript or CSS files. I am not sure if we should
exclude them by default (and allow to include them through an option),
or add an option to explicitly exclude them. I chose the latter.

This filtering should probably be implemented in Hound itself (on the
server). But it does not exist yet.