Don't depend on the `requests` package
`import requests` takes time. It is a known fact, there are many
reports about that. I did not dig too much but it seems to depend on
various factors: installed packages, Python version, etc.

On my computer, it takes around 100-200ms. It's dwarfed by the network
time of a request to our Hound server (that takes around 500ms for
simple queries with only a few results [*]). But it is not negligible
and may even be higher on other computers. Using `urllib` is simple
here and the code is clear, even for humans. So let's use that
instead.

[*] That looks quite high to me. I'll have to dig into that as well.